contracts moved to different directory
