+++ src/Connection/ConnectionInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Anik\Amqp\Contracts;
+namespace Anik\Amqp\Connection;
 
 use PhpAmqpLib\Connection\AbstractConnection;
 
@@ -8,7 +8,7 @@ interface ConnectionInterface
 {
     public function getConnection(): AbstractConnection;
 
-    public function getChannel(): ChannelInterface;
+    public function getChannel(?int $channelId): ChannelInterface;
 
     public function close(): void;
 }
